feat-#146: refactor health module
