+++ src/health/controllers/health.controller.spec.ts
@@ -1,0 +1,24 @@
+import { ConfigService } from '@nestjs/config';
+import { TerminusModule } from '@nestjs/terminus';
+import { Test, TestingModule } from '@nestjs/testing';
+
+import { HealthController } from './health.controller';
+import { HealthService } from '../services';
+
+describe('HealthController', () => {
+  let controller: HealthController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      imports: [TerminusModule],
+      controllers: [HealthController],
+      providers: [HealthService, ConfigService],
+    }).compile();
+
+    controller = module.get<HealthController>(HealthController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
