+++ src/health/health.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TerminusModule } from '@nestjs/terminus';
-import { HealthController } from './health.controller';
-import { HealthService } from './health.service';
+
+import { HealthController } from './controllers';
+import { HealthService } from './services';
 
 @Module({
   imports: [TerminusModule],
