--- src/health/health.service.spec.ts
@@ -1,21 +1,0 @@
-import { ConfigService } from '@nestjs/config';
-import { HealthCheckService, TerminusModule } from '@nestjs/terminus';
-import { Test, TestingModule } from '@nestjs/testing';
-import { HealthService } from './health.service';
-
-describe('HealthService', () => {
-  let service: HealthService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      imports: [TerminusModule],
-      providers: [HealthService, ConfigService],
-    }).compile();
-
-    service = module.get<HealthService>(HealthService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
