+++ src/health/services/health.service.spec.ts
@@ -1,23 +1,22 @@
 import { ConfigService } from '@nestjs/config';
 import { TerminusModule } from '@nestjs/terminus';
 import { Test, TestingModule } from '@nestjs/testing';
-import { HealthController } from './health.controller';
+
 import { HealthService } from './health.service';
 
-describe('HealthController', () => {
-  let controller: HealthController;
+describe('HealthService', () => {
+  let service: HealthService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [TerminusModule],
-      controllers: [HealthController],
       providers: [HealthService, ConfigService],
     }).compile();
 
-    controller = module.get<HealthController>(HealthController);
+    service = module.get<HealthService>(HealthService);
   });
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
+    expect(service).toBeDefined();
   });
 });
